+++ src/main/java/com/infnet/avaliacao/controller/util/PathConstant.java
@@ -4,6 +4,6 @@ public final class PathConstant {
 
     public static final String PATH_USUARIO = "/cadastro/usuario";
     public static final String PATH_TEMPLATE_AVALIACAO = "/template/avaliacao";
-    public static final String PATH_TEMPLATE_PERGUNTA = "/template/avaliacao/topico";
+    public static final String PATH_TEMPLATE_TOPICO = "/template/avaliacao/topico";
 
 }
