+++ src/main/java/com/infnet/avaliacao/dto/impl/TemplateAvaliacaoTopicoPerguntaDTO.java
@@ -86,14 +86,17 @@ public class TemplateAvaliacaoTopicoPerguntaDTO implements IDTO<TemplateAvaliaca
      * esta classe associativa tem 3 atributos e o hibernate tem dificuldade de gerenciar essa quantidade de atributos então a associacao ocorre
      * na mao mesmo.
      * @param templatePerguntaDTOList templatePerguntaDTOList
+     * @param templateTopicoDTO templateTopicoDTO
      * @return List<TemplateAvaliacaoTopicoPerguntaDTO>
      */
-    public static List<TemplateAvaliacaoTopicoPerguntaDTO> produceAssociativeClass(List<TemplatePerguntaDTO> templatePerguntaDTOList){
+    public static List<TemplateAvaliacaoTopicoPerguntaDTO> produceAssociativeClass(List<TemplatePerguntaDTO> templatePerguntaDTOList,
+                                                                                   TemplateTopicoDTO templateTopicoDTO){
         TemplateAvaliacaoTopicoPerguntaDTO templateAvaliacaoTopicoPerguntaDTO;
         List<TemplateAvaliacaoTopicoPerguntaDTO> templateAvaliacaoTopicoPerguntaDTOList = new ArrayList<>();
         for(TemplatePerguntaDTO templatePerguntaDTO : templatePerguntaDTOList){
             templateAvaliacaoTopicoPerguntaDTO = new TemplateAvaliacaoTopicoPerguntaDTO();
             templateAvaliacaoTopicoPerguntaDTO.setTemplatePergunta(templatePerguntaDTO.toEntity());
+            templateAvaliacaoTopicoPerguntaDTO.setTemplateTopico(templateTopicoDTO.toEntity());
             templateAvaliacaoTopicoPerguntaDTOList.add(templateAvaliacaoTopicoPerguntaDTO);
         }
         return templateAvaliacaoTopicoPerguntaDTOList;
