+++ src/main/java/com/infnet/avaliacao/entity/TemplateAvaliacaoTopicoPergunta.java
@@ -20,15 +20,15 @@ public class TemplateAvaliacaoTopicoPergunta implements Serializable {
     @Column(name = "id", nullable = false)
     private Long id;
 
-    @ManyToOne(cascade = CascadeType.ALL)
+    @ManyToOne(cascade = {CascadeType.PERSIST, CascadeType.REFRESH, CascadeType.REMOVE})
     @JoinColumn(name = "id_template_avaliacao", nullable = false)
     private TemplateAvaliacao templateAvaliacao;
 
-    @ManyToOne(cascade = CascadeType.ALL)
+    @ManyToOne(cascade = {CascadeType.PERSIST, CascadeType.REFRESH, CascadeType.REMOVE})
     @JoinColumn(name = "id_template_topico", nullable = false)
     private TemplateTopico templateTopico;
 
-    @ManyToOne(cascade = CascadeType.ALL)
+    @ManyToOne(cascade = {CascadeType.PERSIST, CascadeType.REFRESH, CascadeType.REMOVE})
     @JoinColumn(name = "id_template_pergunta", nullable = false)
     private TemplatePergunta templatePergunta;
