+++ src/main/resources/templates/template/avaliacao/topico/form.html
@@ -35,7 +35,7 @@
 
                         <h2>
                             <span th:text="#{template.avaliacao.montar.avaliacao.curso}"></span>
-                            <!--: <span th:text="${templatePerguntaDTO.titulo}"></span>-->
+                            <!--: <span th:text="${templateTopicoDTO.titulo}"></span>-->
                         </h2>
 
                         <hr />
@@ -46,9 +46,10 @@
                         </div>
 
                         <div class="col-lg-12"> &nbsp; </div>
-                        <form method="POST" th:action="@{/template/avaliacao/topico/save}" th:object="${templatePerguntaDTO}" id="form">
+                        <form method="POST" th:action="@{/template/avaliacao/topico/save}" th:object="${templateTopicoDTO}" id="form">
                             <div class="form-group">
                                 <input type="hidden" th:field="*{id}" id="id"/>
+                                <input type="hidden" th:field="*{enunciado}"/>
                                 <table class="table table-striped table-hover table-bordered">
                                     <thead>
                                         <tr>
@@ -65,7 +66,7 @@
                                             <td th:text="${templatePergunta.questao}"></td>
                                             <td>
                                                  <input type="checkbox" th:value="${templatePergunta.id}" name="checks[]"
-                                                        th:checked="${#lists.contains(templatePerguntaDTO.idsTemplatePerguntaSelecionados, templatePergunta.id)}"
+                                                        th:checked="${#lists.contains(templateTopicoDTO.idsTemplatePerguntaSelecionados, templatePergunta.id)}"
                                                         onclick="selecionarPergunta()"/>
                                             </td>
                                             <td>
